cpu: impl ins + rework internals to use 32bit values

+++ debugger/src/breakpoints.rs
@@ -4,7 +4,7 @@ mod breakpoints_test;
 
 #[derive(Default)]
 pub struct Breakpoints {
-    breakpoints: Vec<usize>,
+    breakpoints: Vec<u32>,
 }
 
 // a list of addresses for the debugger to break on when CS:IP reach one of them
@@ -15,7 +15,7 @@ impl Breakpoints {
         }
     }
 
-    pub fn add(&mut self, bp: usize) -> Option<usize> {
+    pub fn add(&mut self, bp: u32) -> Option<u32> {
         if self.breakpoints.iter().find(|&&x|x == bp).is_none() {
             self.breakpoints.push(bp);
             Some(bp)
@@ -24,7 +24,7 @@ impl Breakpoints {
         }
     }
 
-    pub fn remove(&mut self, bp: usize) -> Option<usize> {
+    pub fn remove(&mut self, bp: u32) -> Option<u32> {
         // TODO later: simplify when https://github.com/rust-lang/rust/issues/40062 is stable
         match self.breakpoints.iter().position(|x| *x == bp) {
             Some(pos) => {
@@ -36,7 +36,7 @@ impl Breakpoints {
     }
 
     // returns a Vec with breakpoints sorted ascending
-    pub fn get(&self) -> Vec<usize> {
+    pub fn get(&self) -> Vec<u32> {
         let mut sorted = self.breakpoints.clone();
         sorted.sort();
         sorted
@@ -47,7 +47,7 @@ impl Breakpoints {
     }
 
     // returns true if address is at breakpoint
-    pub fn hit(&self, address: usize) -> bool {
+    pub fn hit(&self, address: u32) -> bool {
         self.breakpoints.iter().any(|&x| x == address)
     }
 }
